+++ application/src/main/kotlin/com/piotr/marketbroker/domain/accountdetails/AccountDetailsOpeningOrdersAttribute.kt
@@ -78,7 +78,7 @@ data class OpeningOrdersRecord (
     val quoteMode: String,
 
     @JsonProperty("Stake")
-    val stake: Int,
+    val stake: Float,
 
     @JsonProperty("Status")
     val status: Int,
